Split base_setup tests by feature
